Renamed sites.py as utils.py
